Create concrete handlers for Insight, Motion and Socket/Light Switch.
No logic/code changes yet, just extraction/separation.

Fixes #12105
